Remove propTypes.children from PageEndCTA
It's possible for props.children to be an array instead of an object, e.g. if multiple elements w/ no parent are provided; react currently complains about this w/ propTypes defined. I don't think there's a way to specify that a prop contain either an object *or* an array, so we should probably just remove it from propTypes.