Revert "add AccessibleAlert component."
This reverts commit 8720213ff825d00a293803fe1b66323fff5bc3a4.

The AccessibleAlert component makes things work with VoiceOver in
OS X Safari and Chrome, but then *breaks* things in NVDA because
it keeps announcing the empty alert instead of the dialog contents.

It's easier to just keep the code as simple as possible and rely on
UA vendors to make role="alert" work in a consistent way that doesn't
drive developers insane.